+++ src/django_admin_select_filter/__init__.py
@@ -3,6 +3,7 @@ from django_admin_select_filter.filters import (
     ChoiceFilter,
     ForeignKeyFilter,
 )
+from django_admin_select_filter.views import Select2FilterOptionsView
 
 __version__ = "0.1.0"
 
@@ -10,4 +11,5 @@ __all__ = [
     "BaseSelectFilter",
     "ChoiceFilter",
     "ForeignKeyFilter",
+    "Select2FilterOptionsView",
 ]