+++ src/django_admin_select_filter/constants.py
@@ -1,0 +1,3 @@
+from typing import Final
+
+DEFAULT_ASYNC_CALL_URL: Final[str] = "django_admin_select_filter/options/"